Move XCM staff to xcm_config.rs
